Dissolve Move namespace into Activities & Traits
