+++ OpenRA.Mods.D2k/Activities/SwallowActor.cs
@@ -14,7 +14,8 @@ using OpenRA.Activities;
 using OpenRA.GameRules;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.D2k.Traits;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Activities;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.D2k.Activities
