+++ OpenRA.Mods.D2k/Traits/Render/WithProductionOverlay.cs
@@ -14,6 +14,7 @@ using OpenRA.Graphics;
 using OpenRA.Mods.Common.Traits;
 using OpenRA.Mods.RA;
 using OpenRA.Mods.RA.Buildings;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.D2k.Traits
