+++ OpenRA.Mods.RA/AI/AttackOrFleeFuzzy.cs
@@ -12,7 +12,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using AI.Fuzzy.Library;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.GameRules;
 using OpenRA.Traits;
