+++ OpenRA.Mods.RA/Activities/Attack.cs
@@ -9,7 +9,6 @@
 #endregion
 
 using OpenRA.Activities;
-using OpenRA.Mods.RA.Move;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
