+++ OpenRA.Mods.RA/Activities/AttackMoveActivity.cs
@@ -10,7 +10,6 @@
 
 using System.Collections.Generic;
 using OpenRA.Activities;
-using OpenRA.Mods.RA.Move;
 using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
