+++ OpenRA.Mods.RA/Activities/ExternalCaptureActor.cs
@@ -10,7 +10,7 @@
 
 using OpenRA.Activities;
 using OpenRA.Effects;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA.Activities
