+++ OpenRA.Mods.RA/Activities/Leap.cs
@@ -12,7 +12,7 @@ using System;
 using System.Linq;
 using OpenRA.Activities;
 using OpenRA.GameRules;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Mods.RA.Render;
 using OpenRA.Traits;
