+++ OpenRA.Mods.RA/Attack/AttackFollow.cs
@@ -10,7 +10,7 @@
 
 using System;
 using OpenRA.Activities;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Traits;
 
 namespace OpenRA.Mods.RA
