+++ OpenRA.Mods.RA/Crates/GiveUnitCrateAction.cs
@@ -10,7 +10,7 @@
 
 using System.Collections.Generic;
 using System.Linq;
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Primitives;
 using OpenRA.Traits;
