+++ OpenRA.Mods.RA/Crushable.cs
@@ -10,10 +10,9 @@
 
 using System.Linq;
 using OpenRA.Mods.Common.Traits;
-using OpenRA.Mods.RA.Move;
 using OpenRA.Traits;
 
-namespace OpenRA.Mods.RA
+namespace OpenRA.Mods.RA.Traits
 {
 	[Desc("This actor is crushable.")]
 	class CrushableInfo : ITraitInfo
