+++ OpenRA.Mods.RA/Scripting/Properties/MobileProperties.cs
@@ -8,7 +8,8 @@
  */
 #endregion
 
-using OpenRA.Mods.RA.Move;
+using OpenRA.Mods.RA.Activities;
+using OpenRA.Mods.RA.Traits;
 using OpenRA.Scripting;
 using OpenRA.Traits;
 
@@ -30,14 +31,14 @@ namespace OpenRA.Mods.RA.Scripting
 			"(in cells) that will be considered close enough to complete the activity.")]
 		public void Move(CPos cell, int closeEnough = 0)
 		{
-			self.QueueActivity(new Move.Move(self, cell, WRange.FromCells(closeEnough)));
+			self.QueueActivity(new Move(self, cell, WRange.FromCells(closeEnough)));
 		}
 
 		[ScriptActorPropertyActivity]
 		[Desc("Moves within the cell grid, ignoring lane biases.")]
 		public void ScriptedMove(CPos cell)
 		{
-			self.QueueActivity(new Move.Move(self, cell));
+			self.QueueActivity(new Move(self, cell));
 		}
 
 		[ScriptActorPropertyActivity]
